+++ src/components/InRequest/InRequestEditPanel.tsx
@@ -45,7 +45,10 @@ import {
 } from "@fluentui/react-icons";
 import { SENSITIVITY_CODES } from "constants/SensitivityCodes";
 import { UserContext } from "providers/UserProvider";
-import { useAdditionalInfo, useUpdateAdditionalInfo } from "api/AdditionalInfoApi";
+import {
+  useAdditionalInfo,
+  useUpdateAdditionalInfo,
+} from "api/AdditionalInfoApi";
 
 /* FluentUI Styling */
 const useStyles = makeStyles({
@@ -170,11 +173,17 @@ export const InRequestEditPanel: FunctionComponent<IInRequestEditPanel> = (
     const data2 = { ...data, MPCN: mpcn, SSN: undefined } as IInRequest;
 
     // Is the SSN different than what came in and therefore needs pushed to the server
-    if(data.SSN !== undefined && data.SSN !== "" && data.SSN !== additionalInfo.data?.[0]?.Title)
-    {
-      updateAdditionalInfo.mutate({Id: additionalInfo.data?.[0]?.Id ?? -1, Title:data.SSN, RequestId: data.Id})
+    if (
+      data.SSN !== undefined &&
+      data.SSN !== "" &&
+      data.SSN !== additionalInfo.data?.[0]?.Title
+    ) {
+      updateAdditionalInfo.mutate({
+        Id: additionalInfo.data?.[0]?.Id ?? -1,
+        Title: data.SSN,
+        RequestId: data.Id,
+      });
     }
-
 
     updateRequest.mutate(data2, {
       onSuccess: () => {
@@ -376,7 +385,9 @@ export const InRequestEditPanel: FunctionComponent<IInRequestEditPanel> = (
                     control={control}
                     defaultValue={additionalInfo?.data?.[0]?.Title ?? ""}
                     rules={{
-                      required: additionalInfo.data?.[0] ? "SSN is required" : "",
+                      required: additionalInfo.data?.[0]
+                        ? "SSN is required"
+                        : "",
                       minLength: {
                         value: 9,
                         message: "SSN cannot be less than 9 digits",
@@ -436,18 +447,118 @@ export const InRequestEditPanel: FunctionComponent<IInRequestEditPanel> = (
                     }
                   </Text>
                 )}
-                {additionalInfo?.data?.length === 0 && <Text
-                  weight="regular"
-                  size={200}
-                  className={classes.fieldDescription}
-                >
-                  NOTE: You did not enter the original SSN, and therefore cannot
-                  view it. If it was inccorect and needs updated, you may enter
-                  the full SSN to overwrite the original.
-                </Text>
-                }
+                {additionalInfo?.data?.length === 0 && (
+                  <Text
+                    weight="regular"
+                    size={200}
+                    className={classes.fieldDescription}
+                  >
+                    NOTE: You did not enter the original SSN, and therefore
+                    cannot view it. If it was inccorect and needs updated, you
+                    may enter the full SSN to overwrite the original.
+                  </Text>
+                )}
               </div>
             )}
+            <div className={classes.fieldContainer}>
+              <Label
+                htmlFor="jobTitleId"
+                size="small"
+                weight="semibold"
+                className={classes.fieldLabel}
+                required
+              >
+                <NumberFieldIcon className={classes.fieldIcon} />
+                Job/Duty Title
+              </Label>
+              <Controller
+                name="jobTitle"
+                control={control}
+                defaultValue={""}
+                rules={{
+                  required: "Job/Duty Title is required",
+                  maxLength: {
+                    value: 100,
+                    message:
+                      "Job/Duty Title cannot be longer than 100 characters",
+                  },
+                }}
+                render={({ field }) => (
+                  <Input
+                    {...field}
+                    aria-describedby="jobTitleErr"
+                    id="jobTitleId"
+                  />
+                )}
+              />
+              {errors.jobTitle && (
+                <Text id="jobTitleErr" className={classes.errorText}>
+                  {errors.jobTitle.message}
+                </Text>
+              )}
+            </div>
+            <div className={classes.fieldContainer}>
+              <Label
+                htmlFor="dutyPhoneId"
+                size="small"
+                weight="semibold"
+                className={classes.fieldLabel}
+                required
+              >
+                <NumberFieldIcon className={classes.fieldIcon} />
+                Duty Phone #
+              </Label>
+              <Controller
+                name="dutyPhone"
+                control={control}
+                defaultValue={""}
+                rules={{
+                  required: "Duty Phone # is required",
+                  pattern: {
+                    value: /^\d{3}-\d{3}-\d{4}$/,
+                    message: "Phone number must be of the form ###-###-####",
+                  },
+                }}
+                render={({ field }) => (
+                  <Input
+                    {...field}
+                    aria-describedby="dutyPhoneErr"
+                    onInput={(e) => {
+                      let endsDash = false;
+                      if (e.currentTarget.value.match(/-$/)) {
+                        endsDash = true;
+                      }
+                      e.currentTarget.value = e.currentTarget.value.replace(
+                        /\D/g,
+                        ""
+                      );
+                      const size = e.currentTarget.value.length;
+                      if (size > 3 || endsDash) {
+                        // If we have more than 3 numbers, or we have either ###- or ###-###-
+                        // The second condition allows them to type a dash, otherwise the code would "reject" it
+                        e.currentTarget.value =
+                          e.currentTarget.value.slice(0, 3) +
+                          "-" +
+                          e.currentTarget.value.slice(3, 11);
+                      }
+                      if (size > 6 || (size > 5 && endsDash)) {
+                        e.currentTarget.value =
+                          e.currentTarget.value.slice(0, 7) +
+                          "-" +
+                          e.currentTarget.value.slice(7);
+                      }
+                    }}
+                    type="tel"
+                    id="dutyPhoneId"
+                  />
+                )}
+              />
+              {errors.dutyPhone && (
+                <Text id="dutyPhoneErr" className={classes.errorText}>
+                  {errors.dutyPhone.message}
+                </Text>
+              )}
+            </div>
             <div className={classes.fieldContainer}>
               <Label
                 id="gradeRankId"